fixed some stuff and added dotenv
