+++ .gitignore
@@ -1,0 +1,3 @@
+.env
+output
+node_modules
